+++ changelog.md
@@ -3,6 +3,7 @@
 ## What's Changed
 
 - Avoid potential memory leaks
+- Clear in-memory caches on stop
 
 # Version 2025.11.27 (2025-11-25)